requirements.txt
boto3>=1.34.0
botocore>=1.34.0
datasets>=2.18.0
pandas>=2.0.0
numpy>=1.26.0
scikit-learn>=1.4.0
xgboost>=2.0.0
shap>=0.44.0
streamlit>=1.32.0
plotly>=5.20.0
matplotlib>=3.8.0
seaborn>=0.13.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
tqdm>=4.66.0
nbformat>=5.9.0
ipykernel>=6.29.0
